+++ packages/npm/knockback-core.js
@@ -109,7 +109,7 @@ kb = (function() {
     });
   };
 
-  kb.renderAutoReleasedTemplate = function(template, view_model, options) {
+  kb.renderTemplate = function(template, view_model, options) {
     var el, observable;
     if (options == null) {
       options = {};
@@ -122,6 +122,14 @@ kb = (function() {
     kb.releaseOnNodeRemove(view_model, el);
     observable.dispose();
     return el;
+  };
+
+  kb.renderAutoReleasedTemplate = function(template, view_model, options) {
+    if (options == null) {
+      options = {};
+    }
+    legacyWarning('kb.renderAutoReleasedTemplate', '0.16.2', 'Please use kb.renderTemplate instead');
+    return this.renderTemplate(template, view_model, options = {});
   };
 
   kb.applyBindings = function(view_model, node) {
@@ -1688,29 +1696,40 @@ kb.sortedIndexWrapAttr = kb.siwa = function(attribute_name, wrapper_constructor)
 */
 
 
-ko.bindingHandlers['kb-inject'] = {
+ko.bindingHandlers['inject'] = {
   'init': function(element, value_accessor, all_bindings_accessor, view_model) {
-    var data;
+    var data, wrapper;
     data = ko.utils.unwrapObservable(value_accessor());
-    return ko.computed(function() {
+    wrapper = ko.dependentObservable(function() {
+      var key, value, _results;
       if (_.isFunction(data)) {
         return data(view_model, element, value_accessor, all_bindings_accessor);
       } else if (_.isObject(data)) {
-        return _.extend(view_model, data);
+        _results = [];
+        for (key in data) {
+          value = data[key];
+          if (_.isObject(value) && value.resolve && _.isFunction(value.resolve)) {
+            _results.push(view_model[key] = value.resolve(view_model, element, value_accessor, all_bindings_accessor));
+          } else {
+            _results.push(view_model[key] = value);
+          }
+        }
+        return _results;
       }
     });
+    return wrapper.dispose();
   }
 };
 
 kb.injectApps = function(root) {
-  var app_el, app_els, getAppElements, _i, _len;
-  app_els = [];
+  var app, apps, getAppElements, options, _i, _len;
+  apps = [];
   getAppElements = function(el) {
-    var child_el, _i, _len, _ref;
-    if (el.attributes && _.find(el.attributes, function(attr) {
+    var attr, child_el, _i, _len, _ref;
+    if (el.attributes && (attr = _.find(el.attributes, function(attr) {
       return attr.name === 'kb-app';
-    })) {
-      app_els.push(el);
+    }))) {
+      apps.push([el, attr]);
     }
     _ref = el.childNodes;
     for (_i = 0, _len = _ref.length; _i < _len; _i++) {
@@ -1719,9 +1738,20 @@ kb.injectApps = function(root) {
     }
   };
   getAppElements(root || document);
-  for (_i = 0, _len = app_els.length; _i < _len; _i++) {
-    app_el = app_els[_i];
-    kb.applyBindings({}, app_el);
+  for (_i = 0, _len = apps.length; _i < _len; _i++) {
+    app = apps[_i];
+    if (app[1].value) {
+      options = ko.utils.buildEvalWithinScopeFunction("{" + app[1].value + "}", 0)();
+    }
+    options || (options = {});
+    options.view_model || (options.view_model = {});
+    if (options.beforeBinding) {
+      options.beforeBinding(options.view_model, app[0], options);
+    }
+    kb.applyBindings(options.view_model, app[0], {});
+    if (options.afterBinding) {
+      options.afterBinding(options.view_model, app[0], options);
+    }
   }
 };